socket-json-postprocessor-python-image-example/requirements.txt
pyinstaller
requests
sysv_ipc
pillow